Renamed utils to common. Can think of a better name later.
