+++ app/upgrades/v1.17.0/common/ibc_marker.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"fmt"
