+++ app/upgrades/v1.17.0/rc3/upgrades.go
@@ -5,7 +5,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	provenance "github.com/provenance-io/provenance/app"
 	"github.com/provenance-io/provenance/app/upgrades"
-	"github.com/provenance-io/provenance/app/upgrades/v1.17.0/utils"
+	"github.com/provenance-io/provenance/app/upgrades/v1.17.0/common"
 )
 
 func UpgradeStrategy(ctx sdk.Context, app *provenance.App, vm module.VersionMap) (module.VersionMap, error) {
@@ -15,6 +15,6 @@ func UpgradeStrategy(ctx sdk.Context, app *provenance.App, vm module.VersionMap)
 		return nil, err
 	}
 
-	utils.UpdateIbcMarkerDenomMetadata(ctx, app)
+	common.UpdateIbcMarkerDenomMetadata(ctx, app)
 	return newVM, err
 }
